refactor: move arguments and options from parameters module
